Paralellization of the code and bugs fixed.
As the potential calculation is a really intensive process,
now it can be done in several cores. Right now the code
automatically creates two processes for this purpose, but I
intend to make the number of processes an optional parameter
in a future commit.

Also, some bugs and improvements were made in the temperature
calculation for the gas.