Previous temperature calculation method discarded.
It was simply wrong, from a theoretical point of view,
although the reasons are not so clear to me. There doesn't
seem to be a way of escaping from the iterative method
commonly found in the liperature. I switched to a simpler
approach: generating an isothermal disk, with appropriate
temperature, and running the disk for a while to let it
relax.